Added `keep_passed` nextflow parameter.
The `keep_passed` nextflow parameter takes a value of `true` or `false` and is set to a default value of `true`.  When set to `false` and module test result is a `PASSED`, then the `test` directory is deleted to save space.  When `keep_passed` is set to `true`, the `test` directory will be retained no matter if the module test `PASSED` or `FAILED`.